+++ notification/slack/usergroupsender.go
@@ -4,14 +4,11 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"net/url"
 	"strings"
-	"text/template"
 
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	"github.com/slack-go/slack"
-	"github.com/slack-go/slack/slackutilsx"
 	"github.com/target/goalert/config"
 	"github.com/target/goalert/notification"
 	"github.com/target/goalert/permission"
@@ -19,16 +16,19 @@ import (
 	"github.com/target/goalert/util/log"
 )
 
+// UserGroupSender processes on-call notifications by updating the members of a Slack user group.
 type UserGroupSender struct {
 	*ChannelSender
 }
 
 var _ notification.Sender = (*UserGroupSender)(nil)
 
+// UserGroupSender returns a new UserGroupSender wrapping the given ChannelSender.
 func (s *ChannelSender) UserGroupSender() *UserGroupSender {
 	return &UserGroupSender{s}
 }
 
+// Send implements notification.Sender.
 func (s *UserGroupSender) Send(ctx context.Context, msg notification.Message) (*notification.SentMessage, error) {
 	err := permission.LimitCheckAny(ctx, permission.User, permission.System)
 	if err != nil {
@@ -78,6 +78,7 @@ func (s *UserGroupSender) Send(ctx context.Context, msg notification.Message) (*
 	ugID, chanID, _ := strings.Cut(t.Dest.Value, ":")
 	cfg := config.FromContext(ctx)
 
+	// If any users are missing, we need to abort and let the channel know.
 	if len(missing) > 0 {
 		// TODO: add link action button to invite missing users
 		var buf bytes.Buffer
@@ -103,6 +104,9 @@ func (s *UserGroupSender) Send(ctx context.Context, msg notification.Message) (*
 		return &notification.SentMessage{State: notification.StateSent, StateDetails: "missing users, sent error to channel"}, nil
 	}
 
+	// If no users are on-call, we need to abort and let the channel know.
+	//
+	// This is because we can't update the user group with no members.
 	if len(slackUsers) == 0 {
 		var buf bytes.Buffer
 		err := userGroupErrorEmpty.Execute(&buf, userGroupError{
@@ -129,8 +133,14 @@ func (s *UserGroupSender) Send(ctx context.Context, msg notification.Message) (*
 
 	err = s.withClient(ctx, func(c *slack.Client) error {
 		_, err := c.UpdateUserGroupMembersContext(ctx, ugID, strings.Join(slackUsers, ","))
-		return err
+		if err != nil {
+			return fmt.Errorf("update user group '%s': %w", ugID, err)
+		}
+
+		return nil
 	})
+
+	// If there was an error, we need to abort and let the channel know.
 	if err != nil {
 		errID := uuid.New()
 		log.Log(log.WithField(ctx, "SlackUGErrorID", errID), err)
@@ -157,63 +167,3 @@ func (s *UserGroupSender) Send(ctx context.Context, msg notification.Message) (*
 
 	return &notification.SentMessage{State: notification.StateDelivered}, nil
 }
-
-type userGroupError struct {
-	ErrorID      uuid.UUID
-	GroupID      string
-	ScheduleID   string
-	ScheduleName string
-	Missing      []notification.User
-	Linked       []string
-
-	callbackFunc func(string, ...url.Values) string
-}
-
-func (e userGroupError) ErrorRef() string {
-	return fmt.Sprintf("`SlackUGErrorID=%s`", e.ErrorID.String())
-}
-
-func (e userGroupError) GroupRef() string {
-	return fmt.Sprintf("<!subteam^%s>", e.GroupID)
-}
-
-func (e userGroupError) MissingUserRefs() string {
-	var refs []string
-	for _, u := range e.Missing {
-		urlStr := e.callbackFunc(fmt.Sprintf("users/%s", url.PathEscape(u.ID)))
-		refs = append(refs, slackLink(urlStr, u.Name))
-	}
-	return strings.Join(refs, ", ")
-}
-
-func (e userGroupError) LinkedUserRefs() string {
-	var refs []string
-	for _, u := range e.Linked {
-		refs = append(refs, fmt.Sprintf("<@%s>", u))
-	}
-	return strings.Join(refs, " ")
-}
-
-func (e userGroupError) ScheduleRef() string {
-	urlStr := e.callbackFunc("schedules/" + url.PathEscape(e.ScheduleID))
-	return slackLink(urlStr, e.ScheduleName)
-}
-
-func slackLink(url, label string) string {
-	return fmt.Sprintf("<%s|%s>", slackutilsx.EscapeMessage(url), slackutilsx.EscapeMessage(label))
-}
-
-var userGroupErrorMissing = template.Must(template.New("userGroupErrorMissing").Parse(`Hey everyone! I couldn't update {{.GroupRef}} because I couldn't find the following user(s) in Slack: {{.MissingUserRefs}}
-
-If you could have them click the button below to connect their Slack account, that would be great! Hopefully I'll be able to update the user-group next time.
-
-{{.LinkedUserRefs}}`))
-
-var userGroupErrorEmpty = template.Must(template.New("userGroupErrorMissing").Parse(`Hey everyone! I couldn't update {{.GroupRef}} because there is nobody on-call for {{.ScheduleRef}}.
-
-Since a Slack user-group cannot be empty, I'm going to leave it as-is for now.`))
-
-var userGroupErrorUpdate = template.Must(template.New("userGroupErrorMissing").Parse(`Hey everyone! I couldn't update {{.GroupRef}} because I ran into a problem. Maybe touch base with the GoAlert admin(s) to see if they can help? I'm sorry for the inconvenience!
-
-Here's the ID I left with the error in my logs so they can find it:
-{{.ErrorRef}}`))
